@@ -3,14 +3,12 @@ package com.epam.esm.giftcertificate.service;
 import com.epam.esm.exceptionhandler.exceptions.NoSuchItemException;
 import com.epam.esm.exceptionhandler.exceptions.ObjectAlreadyExistsException;
 import com.epam.esm.exceptionhandler.exceptions.ObjectIsInvalidException;
+import com.epam.esm.giftcertificate.direction.DirectionEnum;
 import com.epam.esm.giftcertificate.model.GiftCertificate;
 import com.epam.esm.giftcertificate.repository.GiftCertificateRepository;
 import com.epam.esm.tag.model.Tag;
 import com.epam.esm.tag.service.TagService;
-import com.epam.esm.taggiftcertificate.repository.TagGiftCertificateRepository;
-import com.epam.esm.taggiftcertificate.service.TagGiftCertificateService;
 import com.epam.esm.utils.datavalidation.ParamsValidation;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -24,12 +22,16 @@ import java.util.stream.Collectors;
 public class GiftCertificateService {
     private final GiftCertificateRepository giftCertificateRepository;
     private final TagService tagService;
-    private final TagGiftCertificateService tagGiftCertificateService;
 
-    public GiftCertificateService(GiftCertificateRepository giftCertificateRepository, TagService tagService, TagGiftCertificateRepository tagGiftCertificateRepository, TagGiftCertificateService tagGiftCertificateService) {
+    public static final String DONT_EXIST = "don't exist";
+    public static final String IS_INVALID = " is invalid";
+
+
+    public static final String EMPTY_MESSAGE = "Gift certificates are empty!";
+
+    public GiftCertificateService(GiftCertificateRepository giftCertificateRepository, TagService tagService) {
         this.giftCertificateRepository = giftCertificateRepository;
         this.tagService = tagService;
-        this.tagGiftCertificateService = tagGiftCertificateService;
     }
 
     @Transactional
@@ -44,20 +46,17 @@ public class GiftCertificateService {
         giftCertificateRepository.createGiftCertificate(giftCertificate);
         List<Long> listTagsId = tagService.getTagsIds(giftCertificate.getTags());
         giftCertificateRepository.createTagDependenciesForGiftCertificate(listTagsId, giftCertificateRepository.getGiftCertificatesID(giftCertificate));
-        //get created
+
         return getCertificateById(giftCertificateRepository.getGiftCertificatesID(giftCertificate));
 
     }
 
     public boolean deleteCertificate(long id) {
-
-        if (giftCertificateRepository.deleteGiftCertificate(id)) return true;
-        return false;
-
+        return giftCertificateRepository.deleteGiftCertificate(id);
     }
 
-    public List<GiftCertificate> getAllGiftCertificates() {
-        return giftCertificateRepository.getAllGiftCertificates();
+    public List<GiftCertificate> getAllGiftCertificates(Integer page, Integer size) {
+        return ParamsValidation.isEmptyOrElseThrowPageException(setTagsInCertificates(giftCertificateRepository.getAllGiftCertificates(page, size)));
     }
 
     @Transactional
@@ -65,6 +64,9 @@ public class GiftCertificateService {
         Optional<Map<String, String>> updatingMap = ParamsValidation.isPatchCertificateValid(giftCertificate);
 
         List<Tag> tagsToUpdate = giftCertificate.getTags();
+        if (updatingMap.isEmpty()) {
+            throw new ObjectIsInvalidException("Update params are empty!");
+        }
         if (tagsToUpdate == null) return giftCertificateRepository.updateGiftCertificate(id, updatingMap.get());
         else {
             if (ParamsValidation.isCertificateHaveValidTags(tagsToUpdate) && giftCertificateRepository.updateGiftCertificate(id, updatingMap.get())) {
@@ -72,7 +74,7 @@ public class GiftCertificateService {
                 List<Tag> tagsToDelete;
 
                 List<String> tagsToUpdateNames = tagsToUpdate.stream().map(Tag::getName).collect(Collectors.toList());
-                List<String> alreadyUsedTagNames = alreadyUsedTags.stream().map(Tag::getName).collect(Collectors.toList());
+                List<String> alreadyUsedTagNames = alreadyUsedTags.stream().map(Tag::getName).toList();
                 List<String> tagsToDeleteNames = new ArrayList<>(alreadyUsedTagNames);
                 tagsToDeleteNames.removeAll(tagsToUpdateNames);
                 tagsToUpdateNames.removeAll(alreadyUsedTagNames);
@@ -101,9 +103,65 @@ public class GiftCertificateService {
         if (giftCertificate == null) {
             throw new NoSuchItemException("GiftCertificate with id = " + id + " doesn't exist");
         }
-        giftCertificate.setTags(tagGiftCertificateService.getAllTagsByCertificate(giftCertificate));
+        giftCertificate.setTags(getAllTagsByCertificate(giftCertificate));
         return giftCertificate;
 
     }
 
+    public List<GiftCertificate> setTagsInCertificates(List<GiftCertificate> gcs) {
+        for (GiftCertificate giftCertificate : gcs) {
+            giftCertificate.setTags(getAllTagsByCertificate(giftCertificate));
+        }
+        return gcs;
+    }
+
+    public List<Tag> getAllTagsByCertificate(GiftCertificate gc) {
+        return tagService.getAllTagsByCertificateID(giftCertificateRepository.getGiftCertificatesID(gc));
+    }
+
+    public List<GiftCertificate> getGiftCertificatesByTagName(String tagName, Integer page, Integer size) {
+        if (tagName != null && !tagName.isEmpty()) {
+            if (tagService.isTagWithNameExists(tagName)) {
+                List<GiftCertificate> gc = ParamsValidation.isCertificatesArentEmptyOrElseThrowNoSuchItem(giftCertificateRepository.getGiftCertificatesByTagName(tagName, page, size), "Gift Certificates with tag " + tagName + DONT_EXIST);
+                return setTagsInCertificates(gc);
+            }
+            throw new NoSuchItemException("Tag with name " + tagName + " doesn't exist");
+        }
+        throw new ObjectIsInvalidException("Tag name " + tagName + IS_INVALID);
+    }
+
+    public List<GiftCertificate> getGiftCertificatesByPart(String part, Integer page, Integer size) {
+        if (part != null && !part.isEmpty()) {
+            List<GiftCertificate> certificates = giftCertificateRepository.getGiftCertificatesByPartOfDescription(part, page, size);
+            if (certificates.isEmpty()) {
+                List<GiftCertificate> gc = ParamsValidation.isCertificatesArentEmptyOrElseThrowNoSuchItem(giftCertificateRepository.getGiftCertificatesByPartOfName(part, page, size), "Gift Certificates with part of name -> " + part + DONT_EXIST);
+                return setTagsInCertificates(gc);
+            }
+            List<GiftCertificate> gc = ParamsValidation.isCertificatesArentEmptyOrElseThrowNoSuchItem(certificates, "Gift Certificates with part of description -> " + part + DONT_EXIST);
+            return setTagsInCertificates(gc);
+        }
+        throw new ObjectIsInvalidException("Part of description -> " + part + IS_INVALID);
+    }
+
+
+    public List<GiftCertificate> getCertificatesSortedByDate(DirectionEnum direction, Integer page, Integer size) {
+        List<GiftCertificate> gc = ParamsValidation.isCertificatesArentEmptyOrElseThrowNoSuchItem(giftCertificateRepository.getCertificatesSortedByDate(direction, page, size), EMPTY_MESSAGE);
+        return setTagsInCertificates(gc);
+    }
+
+
+    public List<GiftCertificate> getCertificatesSortedByName(DirectionEnum direction, Integer page, Integer size) {
+        List<GiftCertificate> gc = ParamsValidation.isCertificatesArentEmptyOrElseThrowNoSuchItem(giftCertificateRepository.getCertificatesSortedByName(direction, page, size), EMPTY_MESSAGE);
+        return setTagsInCertificates(gc);
+    }
+
+
+    public List<GiftCertificate> getCertificatesSortedByDateName(DirectionEnum directionDate, DirectionEnum directionName, Integer page, Integer size) {
+        List<GiftCertificate> gc = ParamsValidation.isCertificatesArentEmptyOrElseThrowNoSuchItem(giftCertificateRepository.getCertificatesSortedByDateName(directionDate, directionName, page, size), EMPTY_MESSAGE);
+        return setTagsInCertificates(gc);
+    }
+
+    public List<GiftCertificate> getCertificatesBySeveralTags(List<Long> tags, Integer page, Integer size) {
+        return ParamsValidation.isEmptyOrElseThrowPageException(setTagsInCertificates(giftCertificateRepository.getCertificatesBySeveralTags(tags, page, size)));
+    }
 }
