@@ -2,7 +2,10 @@ package com.epam.esm.giftcertificate.repository;
 
 import com.epam.esm.giftcertificate.model.GiftCertificate;
 import com.epam.esm.tag.model.Tag;
+import com.epam.esm.giftcertificate.direction.DirectionEnum;
 
+
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -13,7 +16,7 @@ public interface GiftCertificateRepository {
     boolean createTagDependenciesForGiftCertificate(List<Long> tags, long giftCertificateID);
 
 
-    List<GiftCertificate> getAllGiftCertificates();
+    List<GiftCertificate> getAllGiftCertificates(Integer page,Integer size);
 
     GiftCertificate getGiftCertificateByID(long id);
 
@@ -28,4 +31,20 @@ public interface GiftCertificateRepository {
     long getGiftCertificatesID(GiftCertificate giftCertificate);
 
     List<Tag> getAllTagsIdByCertificateId(long id);
+
+    List<GiftCertificate> getGiftCertificatesByTagName(String tagName,Integer page,Integer size);
+
+    List<GiftCertificate> getGiftCertificatesByPartOfName(String part,Integer page,Integer size);
+
+    List<GiftCertificate> getGiftCertificatesByPartOfDescription(String part,Integer page,Integer size);
+
+    List<GiftCertificate> getCertificatesSortedByDate(DirectionEnum direction,Integer page,Integer size);
+
+    List<GiftCertificate> getCertificatesSortedByName(DirectionEnum direction,Integer page,Integer size);
+
+    List<GiftCertificate> getCertificatesSortedByDateName(DirectionEnum directionDate, DirectionEnum directionName,Integer page,Integer size);
+
+    List<GiftCertificate> getCertificatesBySeveralTags(List<Long> tags,Integer page,Integer size);
+
+    BigDecimal getCertificatesPriceByID(long id);
 }
