@@ -1,13 +1,10 @@
 package com.epam.esm.tag.repository;
 
-import com.epam.esm.giftcertificate.repository.GiftCertificateRepository;
 import com.epam.esm.tag.model.Tag;
-import com.epam.esm.utils.AppQuery;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.datasource.embedded.EmbeddedDatabase;
 import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseBuilder;
 import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseType;
@@ -18,6 +15,11 @@ import static org.junit.jupiter.api.Assertions.*;
 
 class TagRepositoryTest {
 
+    public static final String TAG_NAME = "TagName";
+    public static final String TAG_1 = "Tag1";
+    public static final long ID = 1L;
+    public static final long ID1 = 2L;
+    public static final String TAG_2 = "Tag2";
     private TagRepository tagRepository;
     private EmbeddedDatabase embeddedDatabase;
 
@@ -34,23 +36,23 @@ class TagRepositoryTest {
     @Test
     void createTagGetTagByName() {
         Tag tag = new Tag();
-        tag.setName("TagName");
+        tag.setName(TAG_NAME);
         boolean result = tagRepository.createTag(tag);
         assertTrue(result);
-        Tag tagRes = tagRepository.getTagByName("TagName");
+        Tag tagRes = tagRepository.getTagByName(TAG_NAME);
         assertEquals(tag.getName(), tagRes.getName());
     }
 
     @Test
     void getAllTags() {
         Tag tag1 = new Tag();
-        tag1.setName("Tag1");
-        tag1.setId(1L);
+        tag1.setName(TAG_1);
+        tag1.setId(ID);
         Tag tag2 = new Tag();
-        tag2.setName("Tag2");
-        tag2.setId(2L);
+        tag2.setName(TAG_2);
+        tag2.setId(ID1);
         List<Tag> expectedTagList = List.of(tag1,tag2);
-        List<Tag> actualTagList = tagRepository.getAllTags();
+        List<Tag> actualTagList = tagRepository.getAllTags(1,10);
 
         //then
         assertEquals(expectedTagList, actualTagList);
@@ -59,9 +61,9 @@ class TagRepositoryTest {
     @Test
     void getTagByID() {
         Tag tag1 = new Tag();
-        tag1.setName("Tag1");
-        tag1.setId(1L);
-        assertEquals(tag1,tagRepository.getTagByID(1L));
+        tag1.setName(TAG_1);
+        tag1.setId(ID);
+        assertEquals(tag1,tagRepository.getTagByID(ID));
     }
 
 
@@ -69,18 +71,25 @@ class TagRepositoryTest {
     @Test
     void isExistsGetTagsIdAndDeleteTagByID() {
         Tag tag = new Tag();
-        tag.setName("Tag1");
-        assertTrue(tagRepository.isTagExists("Tag1"));
-        assertTrue(tagRepository.deleteTagByID(tagRepository.getTagsID(tag)));
-        assertFalse(tagRepository.isTagExists("Tag1"));
+        tag.setName(TAG_1);
+        assertTrue(tagRepository.isTagExists(TAG_1));
+        assertTrue(tagRepository.deleteTagByID(tagRepository.getTagID(tag)));
+        assertFalse(tagRepository.isTagExists(TAG_1));
     }
 
     @Test
     void getAllTagsByCertificateID() {
         Tag tag = new Tag();
-        tag.setName("Tag1");
-        tag.setId(1L);
-        assertEquals(List.of(tag), tagRepository.getAllTagsByCertificateID(1L));
+        tag.setName(TAG_2);
+        tag.setId(ID1);
+        assertEquals(List.of(tag), tagRepository.getAllTagsByCertificateID(ID1));
+    }
+    @Test
+    void getMostUsedTag(){
+        Tag t = new Tag();
+        t.setId(ID1);
+        t.setName("Tag2");
+        assertEquals(t,tagRepository.getMostUsedTag());
     }
     @AfterEach
     public void drop(){
